requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
websockets
python-jose[cryptography]
passlib[bcrypt]
email-validator